+++ test/unit/mail_2_rmq/parse_email.py
@@ -9,7 +9,6 @@
         test/unit/mail_2_rmq/parse_email.py
 
     Arguments:
-        None
 
 """
 
@@ -32,7 +31,6 @@ sys.path.append(os.getcwd())
 import mail_2_rmq
 import version
 
-# Version
 __version__ = version.__version__
 
 
@@ -41,10 +39,6 @@ class UnitTest(unittest.TestCase):
     """Class:  UnitTest
 
     Description:  Class which is a representation of a unit testing.
-
-    Super-Class:  unittest.TestCase
-
-    Sub-Classes:  None
 
     Methods:
         setUp -> Unit testing initilization.
@@ -59,7 +53,6 @@ class UnitTest(unittest.TestCase):
         Description:  Initialization for unit testing.
 
         Arguments:
-            None
 
         """
 
@@ -69,10 +62,6 @@ class UnitTest(unittest.TestCase):
 
             Description:  Class which is a representation of the email.Parser
                 class.
-
-            Super-Class:  object
-
-            Sub-Classes:  None
 
             Methods:
                 __init__ -> Initialize configuration environment.
@@ -87,7 +76,6 @@ class UnitTest(unittest.TestCase):
                 Description:  Initialization instance of the RQTest class.
 
                 Arguments:
-                        None
 
                 """
 
@@ -100,14 +88,13 @@ class UnitTest(unittest.TestCase):
                 Description:  Stub holder for parsestr method.
 
                 Arguments:
-                        None
 
                 """
 
                 return msg
 
         self.raw_msg = ["Raw", "Email", "Message"]
-        self.PT = ParserTest()
+        self.part = ParserTest()
         self.processed_msg = "RawEmailMessage"
 
     @mock.patch("mail_2_rmq.email.Parser")
@@ -119,11 +106,10 @@ class UnitTest(unittest.TestCase):
         Description:  Test parsing an email.
 
         Arguments:
-            None
 
         """
 
-        mock_parse.Parser.return_value = self.PT
+        mock_parse.Parser.return_value = self.part
         mock_stdin.readlines.return_value = self.raw_msg
 
         self.assertEqual(mail_2_rmq.parse_email(), self.processed_msg)